HID: usbhid: quirks for Corsair devices (K70R, M65RGB, K70RGB, K95RGB)
These devices feature multiple interfaces/endpoints: a legacy BIOS/boot interface (endpoint 0x81), as well as 2 corsair-specific keyboard interfaces (endpoint 0x82, 0x83) and an RGB LED control interface (endpoint 0x84, 0x04)

Because the extra 3 interfaces are not of subclass USB_INTERFACE_SUBCLASS_BOOT, HID_QUIRK_NOGET is not automatically set on them -- we hard-set this quirk on the device to prevent 30+ second report timeouts while initialising the other interfaces.

Additionally the left-side macro keys on the K95RGB generate output on the un-opened 0x82/0x83 endpoints which causes the keyboard to stop responding waiting for this event to be collected.  We enable HID_QUIRK_ALWAYS_POLL to prevent this situation from occurring.